split receipt event handling
